+++ R/deployApp.R
@@ -18,8 +18,8 @@
 #' control to ensure that future you and other collaborators will publish
 #' to the same location.
 #'
-#' @param appDir Directory containing application. Defaults to current working
-#'   directory.
+#' @param appDir A directory containing an application (e.g. a Shiny app
+#'   or plumber API). Defaults to the current directory.
 #' @param appFiles A character vector given relative paths to the files and
 #'   directories to bundle and deploy. The default, `NULL`, will include all
 #'   files in `appDir`, apart from any listed in an `.rscignore` file.
@@ -31,10 +31,8 @@
 #'   parameter indicates the primary one, as a path relative to `appDir`. Can be
 #'   `NULL`, in which case the primary document is inferred from the contents
 #'   being deployed.
-#' @param appSourceDoc If the application is composed of static files (e.g
-#'   HTML), this parameter indicates the source document, if any, as a fully
-#'   qualified path. Deployment information returned by [deployments()] is
-#'   associated with the source document.
+#' @param appSourceDoc `r lifecycle::badge("deprecated")` Please use
+#'   `recordDir` instead.
 #' @param appName Name of application (names must be unique within an account).
 #'   If not supplied on the first deploy, it will be generated from the base
 #'   name of `appDir` and `appTitle`, if supplied. On subsequent deploys,
@@ -152,10 +150,49 @@ deployApp <- function(appDir = getwd(),
                       image = NULL
                       ) {
 
-  condaMode <- FALSE
-
+  check_string(appDir)
+  if (isStaticFile(appDir) && !dirExists(appDir)) {
+    lifecycle::deprecate_warn(
+      when = "0.9.0",
+      what = "deployApp(appDir = 'takes a directory, not a document,')",
+      with = "deployDoc()"
+    )
+    return(deployDoc(
+      appDir,
+      appName = appName,
+      appTitle = appTitle,
+      account = account,
+      server = server,
+      upload = upload,
+      recordDir = recordDir,
+      launch.browser = launch.browser,
+      logLevel = logLevel,
+      lint = lint
+    ))
+  }
   check_directory(appDir)
+  appDir <- normalizePath(appDir)
+
   check_string(appName, allow_null = TRUE)
+
+  if (!is.null(appPrimaryDoc)) {
+    check_string(appPrimaryDoc)
+    if (!file.exists(file.path(appDir, appPrimaryDoc))) {
+      cli::cli_abort("`appPrimaryDoc` not found inside `appDir`")
+    }
+  }
+
+  if (!is.null(appSourceDoc)) {
+    lifecycle::deprecate_warn(
+      when = "0.9.0",
+      what = "deployApp(appSourceDoc)",
+      with = "deployApp(recordDir)",
+    )
+    check_directory(appSourceDoc)
+    recordDir <- appSourceDoc
+  } else if (!is.null(recordDir)) {
+    check_directory(recordDir)
+  }
 
   # set up logging helpers
   logLevel <- match.arg(logLevel)
@@ -190,47 +227,6 @@ deployApp <- function(appDir = getwd(),
     on.exit(options(old_error), add = TRUE)
   }
 
-  # normalize appDir path
-  appDir <- normalizePath(appDir)
-
-  # create the full path that we'll deploy (append document if requested)
-  # TODO(HW): we use appPrimaryDoc here, but we have not inferred it yet
-  # so the appPath will different deneding on whether it's explicitly
-  # supplied or inferred from the files in the directory.
-  appPath <- appDir
-  if (!is.null(appSourceDoc) && nchar(appSourceDoc) > 0) {
-    appPath <- appSourceDoc
-  } else if (!is.null(appPrimaryDoc) && nchar(appPrimaryDoc) > 0) {
-    appPath <- file.path(appPath, appPrimaryDoc)
-    if (!file.exists(appPath)) {
-      stop(appPath, " does not exist")
-    }
-  }
-
-  # if a specific file is named, make sure it's an Rmd or HTML, and just deploy
-  # a single document in this case (this will call back to deployApp with a list
-  # of supporting documents)
-  rmdFile <- ""
-  if (!dirExists(appDir)) {
-    if (grepl("\\.[Rq]md$", appDir, ignore.case = TRUE) ||
-        grepl("\\.html?$", appDir, ignore.case = TRUE)) {
-      return(deployDoc(appDir, appName = appName, appTitle = appTitle,
-                       account = account, server = server, upload = upload,
-                       recordDir = recordDir, launch.browser = launch.browser,
-                       logLevel = logLevel, lint = lint))
-    } else {
-      stop(appDir, " must be a directory, an R Markdown document, or an HTML ",
-           "document.")
-    }
-  }
-
-  # directory for recording deployment
-  if (is.null(recordDir)) {
-    recordDir <- appPath
-  } else {
-    check_directory(recordDir)
-  }
-
   # at verbose log level, generate header
   if (verbose) {
     logger("Deployment log started")
@@ -250,7 +246,8 @@ deployApp <- function(appDir = getwd(),
   }
 
   # determine the deployment target and target account info
-  target <- deploymentTarget(recordDir, appName, appTitle, appId, account, server)
+  recordPath <- findRecordPath(appDir, recordDir, appPrimaryDoc)
+  target <- deploymentTarget(recordPath, appName, appTitle, appId, account, server)
 
   # test for compatibility between account type and publish intent
   if (!isCloudServer(target$server) && identical(upload, FALSE)) {
@@ -268,6 +265,12 @@ deployApp <- function(appDir = getwd(),
   # get the application to deploy (creates a new app on demand)
   withStatus("Preparing to deploy", {
     application <- applicationForTarget(client, accountDetails, target, forceUpdate)
+    saveDeployment(
+      recordPath,
+      target = target,
+      application = application,
+      metadata = metadata
+    )
   })
 
   # Change _visibility_ before uploading data
@@ -282,16 +285,25 @@ deployApp <- function(appDir = getwd(),
   }
 
   if (upload) {
+    python <- getPythonForTarget(python, accountDetails)
+    pythonConfig <- pythonConfigurator(python, forceGeneratePythonEnvironment)
+
     # create, and upload the bundle
     logger("Bundle upload started")
     withStatus(paste0("Uploading bundle (", application$id, ")"), {
-
-      # python is enabled on Connect but not on Shinyapps
-      python <- getPythonForTarget(python, accountDetails)
-      bundlePath <- bundleApp(target$appName, appDir, appFiles,
-                              appPrimaryDoc, contentCategory, verbose, python,
-                              condaMode, forceGeneratePythonEnvironment, quarto,
-                              isCloudServer(accountDetails$server), metadata, image)
+      bundlePath <- bundleApp(
+        appName = target$appName,
+        appDir = appDir,
+        appFiles = appFiles,
+        appPrimaryDoc = appPrimaryDoc,
+        contentCategory = contentCategory,
+        verbose = verbose,
+        pythonConfig = pythonConfig,
+        quarto = quarto,
+        isCloudServer = isCloudServer(accountDetails$server),
+        metadata = metadata,
+        image = image
+      )
 
       if (isCloudServer(accountDetails$server)) {
         bundle <- uploadCloudBundle(client, application$id, bundlePath)
@@ -299,31 +311,17 @@ deployApp <- function(appDir = getwd(),
         bundle <- client$uploadApplication(application$id, bundlePath)
       }
     })
+
+    saveDeployment(
+      recordPath,
+      target = target,
+      application = application,
+      bundleId = bundle$id,
+      metadata = metadata
+    )
   } else {
     # redeploy current bundle
     bundle <- application$deployment$bundle
-  }
-
-  # write a deployment record only if this is the account that owns the content
-  if (is.null(application$owner_username) ||
-      accountDetails$username == application$owner_username) {
-    # save the deployment info for subsequent updates--we do this before
-    # attempting the deployment itself to make retry easy on failure.
-    logger("Saving deployment record for ", target$appName, "-", target$username)
-    saveDeployment(recordDir,
-                   target$appName,
-                   target$appTitle,
-                   target$username,
-                   target$account,
-                   accountDetails$server,
-                   serverInfo(target$server)$hostUrl,
-                   application$id,
-                   bundle$id,
-                   application$url,
-                   metadata)
-  } else {
-    logger("Updating ", target$appName, ", owned by ", application$owner_username,
-        ", from account", accountDetails$username)
   }
 
   if (length(bundle$id) > 0 && nzchar(bundle$id)) {
@@ -364,6 +362,18 @@ deployApp <- function(appDir = getwd(),
   invisible(deploymentSucceeded)
 }
 
+findRecordPath <- function(appDir,
+                           recordDir = NULL,
+                           appPrimaryDoc = NULL) {
+  if (!is.null(recordDir)) {
+    recordDir
+  } else if (!is.null(appPrimaryDoc)) {
+    file.path(appDir, appPrimaryDoc)
+  } else {
+    appDir
+  }
+}
+
 # Shinyapps defaults to public visibility.
 # Other values should be set before data is deployed.
 needsVisibilityChange <- function(server, application, appVisibility = NULL) {
@@ -398,10 +408,17 @@ runDeploymentHook <- function(appDir, option, verbose = FALSE) {
 # deployApp() instead of being exposed to the user. Returns the path to the
 # bundle directory, whereas writeManifest() returns nothing and deletes the
 # bundle directory after writing the manifest.
-bundleApp <- function(appName, appDir, appFiles, appPrimaryDoc,
-                      contentCategory, verbose = FALSE, python = NULL,
-                      condaMode = FALSE, forceGenerate = FALSE, quarto = NULL,
-                      isCloudServer = FALSE, metadata = list(), image = NULL) {
+bundleApp <- function(appName,
+                      appDir,
+                      appFiles,
+                      appPrimaryDoc,
+                      contentCategory = NULL,
+                      verbose = FALSE,
+                      pythonConfig = NULL,
+                      quarto = NULL,
+                      isCloudServer = FALSE,
+                      metadata = list(),
+                      image = NULL) {
   logger <- verboseLogger(verbose)
 
   logger("Inferring App mode and parameters")
@@ -432,21 +449,20 @@ bundleApp <- function(appName, appDir, appFiles, appPrimaryDoc,
   # generate the manifest and write it into the bundle dir
   logger("Generate manifest.json")
   manifest <- createAppManifest(
-      appDir = bundleDir,
-      appMode = appMetadata$appMode,
-      contentCategory = contentCategory,
-      hasParameters = appMetadata$hasParameters,
-      appPrimaryDoc = appMetadata$appPrimaryDoc,
-      users = users,
-      condaMode = condaMode,
-      forceGenerate = forceGenerate,
-      python = python,
-      documentsHavePython = appMetadata$documentsHavePython,
-      retainPackratDirectory = TRUE,
-      quartoInfo = appMetadata$quartoInfo,
-      isCloud = isCloudServer,
-      image = image,
-      verbose = verbose)
+    appDir = bundleDir,
+    appMode = appMetadata$appMode,
+    contentCategory = contentCategory,
+    hasParameters = appMetadata$hasParameters,
+    appPrimaryDoc = appMetadata$appPrimaryDoc,
+    users = users,
+    pythonConfig = pythonConfig,
+    documentsHavePython = appMetadata$documentsHavePython,
+    retainPackratDirectory = TRUE,
+    quartoInfo = appMetadata$quartoInfo,
+    isCloud = isCloudServer,
+    image = image,
+    verbose = verbose
+  )
   manifestJson <- enc2utf8(toJSON(manifest, pretty = TRUE))
   manifestPath <- file.path(bundleDir, "manifest.json")
   writeLines(manifestJson, manifestPath, useBytes = TRUE)
@@ -463,25 +479,6 @@ bundleApp <- function(appName, appDir, appFiles, appPrimaryDoc,
 }
 
 
-getPythonForTarget <- function(path, accountDetails) {
-  # python is enabled on Connect and posit.cloud, but not on Shinyapps
-  targetIsShinyapps <- isShinyappsServer(accountDetails$server)
-  pythonEnabled <- getOption("rsconnect.python.enabled", default = !targetIsShinyapps)
-  if (pythonEnabled) {
-    getPython(path)
-  } else {
-    NULL
-  }
-}
-
-
-# get the record for the application of the given name in the given account, or
-# NULL if no application exists by that name
-getAppByName <- function(client, accountInfo, name) {
-  # NOTE: returns a list with 0 or 1 elements
-  app <- client$listApplications(accountInfo$accountId, filters = list(name = name))
-  if (length(app)) app[[1]] else NULL
-}
 
 # get the record for the application with the given ID in the given account;
 # this isn't used inside the package itself but is invoked from the RStudio IDE
@@ -512,34 +509,49 @@ getAppById <- function(id, account = NULL, server = NULL, hostUrl = NULL) {
 }
 
 applicationForTarget <- function(client, accountInfo, target, forceUpdate) {
-
-  if (is.null(target$appId)) {
-    # list the existing applications for this account and see if we
-    # need to create a new application
-    app <- getAppByName(client, accountInfo, target$appName)
-  } else {
-    # we already know the app's id, so just retrieve the rest of the metadata
+  # Use appId from previous deployment, if it still exists
+  if (!is.null(target$appId)) {
     app <- client$getApplication(target$appId)
+    if (!is.null(app)) {
+      return(app)
+    }
   }
 
-  # if there is no record of deploying this application locally however there
-  # is an application of that name already deployed then confirm
-  if (!is.null(target$appId) && !is.null(app) && interactive() && !forceUpdate) {
-    prompt <- paste("Update application currently deployed at\n", app$url,
-                    "? [Y/n] ", sep = "")
-    input <- readline(prompt)
-    if (nzchar(input) && !identical(input, "y") && !identical(input, "Y"))
-      stop("Application deployment cancelled", call. = FALSE)
+  # Otherwise, see if there's an existing app with this name
+  sameName <- getAppByName(client, accountInfo, target$appName)
+  if (!is.null(sameName)) {
+    # check that it's ok to to use it
+    if (interactive() && !forceUpdate) {
+      cat("\n") # Escape from preparing to deploy line
+      cli::cli_inform(paste0(
+        "There is a currently deployed app with name {.str {target$appName}}",
+        " at {.url {sameName$url}}"
+      ))
+      input <- readline("Do you want to update it? [Y/n] ")
+      if (input %in% c("y", "Y", "")) {
+        return(sameName)
+      }
+
+      cli::cli_abort(c(
+        "Each item of content must have a unique {.arg appName}.",
+        i = "Set {.arg appName} to a new value."
+      ))
+    }
   }
 
-  # create the application if we need to
-  if (is.null(app)) {
-    app <- client$createApplication(target$appName, target$appTitle, "shiny",
-                                    accountInfo$accountId)
-  }
+  # Otherwise, create a new app
+  client$createApplication(
+    target$appName,
+    target$appTitle,
+    "shiny",
+    accountInfo$accountId
+  )
+}
 
-  # return the application
-  app
+getAppByName <- function(client, accountInfo, name) {
+  # NOTE: returns a list with 0 or 1 elements
+  app <- client$listApplications(accountInfo$accountId, filters = list(name = name))
+  if (length(app)) app[[1]] else NULL
 }
 
 validURL <- function(url) {