+++ NEWS.md
@@ -7,6 +7,21 @@
   previously some (like `.Rproj.user` and `"manifest.json"`) were only applied
   to the root directory.
 
+* `deployApp()` is more aggressive about saving deployment data, which should
+  make it less likely that you need to repeat yourself after a failed 
+  deployment. In particular, it now saves both before and after uploading the
+  contents (#677) and it saves when you're updating content originally created
+  by someone else (#270).
+  
+* `deployApp("foo.Rmd")` has been deprecated. It was never documented, and
+  it does the same job as `deployDoc()` (#698).
+
+* `deployApp(appPrimaryDoc)` has been deprecated; it did the same job as 
+  `recordDir`.
+
+* `appDependencies()` now returns an additional column giving the Repository 
+  (#670)
+
 * The `rsconnect.pre.deploy` and `rsconnect.post.deploy` hooks are now always
   called with the content directory, not sometimes the path to a specific file
   (#696).